\name{VignetteExamples}

\alias{EquatorData}
\alias{PolarData}
\alias{equatorLocations}
\alias{equatorValues}
\alias{equatorGrid}
\alias{equatorGridValues}
\alias{polarLocations}
\alias{polarValues}
\alias{polarGrid}
\alias{polarGridValues}
\docType{data}
\title{Data examples for the LatticeKrig Vignette}
\description{
Data sets used for  the R scripts in the vignette
 }
 \usage{
 data(EquatorData)
 data(PolarData)
 }


\format{
 See Section 4 in the {LatticeKrig Vignette} for the use of these data.
 
\strong{EquatorData:}
\code{equatorLocations equatorValues equatorGrid equatorGridValues}

\strong{PolarData:}
\code{polarLocations polarValues polarGrid polarGridValues}

}

\keyword{datasets}
